Remove entrySet() add support
Java 8 added support for streams operating on views, though that intent
never matured. The contract is confusing as there are many opinions of
how it can be implemented. After discussing with Guava and Doug Lea who
had different implementation takes, the least confusing is to not support
this operation.

Added testlib tests for async cache map view and fixed iterator bug that
it detected.