Tests and benchmarks for read buffer implementation strategies
The recent change from a racy to one-off read buffer has a race condition
that may be benign but is detectable during validation. So we need a more
robust way to evaluate strategies. Introduced a test, cleaned out the ad hoc
benchmark, and provided a few approaches. The one-shot buffer was fixed by
using a non-blocking PTL scheme which impacted its performance. The cache's
internal version needs to be updated as well.

The JCTools mpsc array queue offers similar performance to the one-shot queue
but is more consistent (fixed one-shot fluxuates in benchmark). This should be
evaluated in more depth, e.g. may be due to false sharing. It may be best to
rewrite mpsc array queue and adapt to the cache.