Detect gc'd entry in computeIfAbsent
Refactored to optimize large method per jitwatch detection. Added handling
of the entry being collected and recomputed.

Due to the Guava implementation not having an executor, it returns a new value
iff the caller triggered the refresh (therefore computed the new value). If
not, then the old value is returned. Caffeine delegates this work, so since
the spec is not restrictive updated the tests to be aware of the optional
async behavior.