Optimizations to the concurrent stack
Previously the stack lazily removed deleted elements by leaving a bubble if an
interior element was deleted (by remove(Object) or iterator.remove()). This has
a slight disadvantage by increasing the likelyhood that the node resides in
old-gen and by allowing uncontrolled growth of dead nodes if deletion is only
by remove.

The stack now tries to unlink deleted nodes during any traversal and eagerly
sweeps during a clear(). A bubble may be left in place due to concurrent
traversals conflicting. The next field does not need to be volatile due to
being optimistically unlinked and always safe to traverse. Due to the resulting
similarity with Doug Lea's abandoned concurrent stack (no backoff), the node
unlinking logic is modelled to be consistent with his design.