Remove unnecessary store-store memory barriers
Previously was using a store-store relaxed write for lazy visibility, where a
normal store is acceptable. Both rely on the same assumptions that visibility
can be delayed and are ordered by other proceeding memory barriers. Primarily
when a lock is released the store piggy backs on the full memory barrier. In
the case of access/write timestamps, this metadata is best effort enough that
a store-store barrier isn't necessary.