Clean up buffers and padding
The algorithm implementations now use the Buffer interface and are
padded. This makes it easier to experiment with optimizations and
copy code directly into the BoundedBuffer version used by the cache.

The padding field names are updated to be unique instead of shadowing
the inheritted copy. This is stylistic due to previous laziness when
copying padding.