Remove lazy queue in development as no plans on completing
The lazy queue was an experimental mpmc queue that was more
optimistic than CLQ and included a combining arena. The races
were never resolved and, most likely, the queue contract could
not be honored due to the optimistic properties.

Removed @Beta annotation as both `ConcurrentLinkedStack` and
`SingleConsumerQueue` are api stable.