Fix visibility race in ConcurrentLinkedStack
This appears to fail very rarely in the multithreaded test case. The
field should be made visible due to being set prior to publishing, but
oddly it is null (NPEs). This is likely due to reordering which moves
the store to after the transfer despite a memory barrier that should
block that compiler optimization. To enforce ordering the field is now
volatile with a lazySet, which should have minimal impact on read/write
performance.

Adding back `@Beta` annotation because I want more time to let this bake
until I feel confident that there are no more subtle races.