Notify JCache listener in key-order on eviction
The specification requires that listeners are notified in sequence of
key operations on the cache. Like all other jsr107 implementations,
Caffeine was notifying in entry order due to the eviction event being
published asynchronously to the write. A Caffeine CacheWriter is used
to publish the eviction atomically with the cache update, so a JCache
listener should always receive events in the proper sequence.