Fix bad pom being generated (fix #22)
The tracing-api test artifact was being consumed by caffeine and the
tracing-async projects in order to share a test utility. Due to Gradle
generating the Maven pom dependencies in a non-deterministic order, this
resulted in Maven resolving the dependencies incorrectly. Instead of
allowing both a test and compile scope, Maven ignored the compile scope
and did not consume the transitive dependency in downstream projects.

The utility class was moved to caffeine/test and the async usage was
rewritted to be ad hoc.

Verified the pom generated only includes the compile scoped dependency.