Use JVM arg to reliably perform soft reference tests
Soft references are difficult to test due to requiring a major garbage
collection. The typical hacks like filling up the heap don't work
reliably on 64-bit JVMs and each GC algorithm behaves differently. This
meant that we relied on weak references to allude to whether the soft
reference caches worked correctly.

The JVM flag `-XX:SoftRefLRUPolicyMSPerMB=0` instructs the collector to
retain zero soft references (retain for 0 millis per megabyte). This
makes soft references behave as weak references, by being GC'd when
no strong reference exists. The tests can now reliably perform a GC
and assume that the reference was cleared.

Note that running the tests (ReferenceTest) in an IDE requires that
this flag is added as a JVM argument to the launcher. This flag is
added to the build, but may not be honored by the IDE.