Add Bélády's optimal cache to simulator
Unlike other policies, OPT must be run as a post-processing step in
order to evaluate against the full usage history. The amount of work
that could be performed incrementally in parallel with other policies
is miniscule due to terminal entries (17 out of 10M). This is because
the policy does not know if the terminal entry will be seen again and
must wait until the entire trace has been recorded. By careful use of
data structures, the sequential time even in the worse case for a large
trace is only a few seconds.