Early pass of LIRS policy for the simulator
This is a first pass that has a few issues remaining that need to be
addressed. The goal is to gain familiarity with the algorithm so that,
if it was adopted, I have enough of an understanding of to implement it
properly with tests. Secondly the alternative implementations show that
work is needed to find the optimal values for the tuning parameters so
those need to be configurable in the simulator.

This implementation blindly follows the paper. There are a few issues.
First the paper discusses making Lhirs sized at 1%, which isn't done
yet. Second the number of non-resident entries is not capped, as the
paper does not discuss this detail. Third the Jackrabbit optimization
should be an optional configuration so we can see the effect on the
hit rate.