Fix warmup of HIR blocks in LIRS policy
When comparing against the output from the C reference version I discovered
that the warmup for HIR was incorrectly followed. This change results in the
same output for a small sample.

There is still a mistake as verified by running a large sample, multi1, and
seeing that S, Q, evictions, and statistics are askew. While this has a very
small impact on the hit rate, we should still try to find the subtle bug
remaining.

For ease of comparing outputs, the debug format now matches the reference's.