Do not promote HIR if already at the top of the stack
Comparing multi1 trace between this version and the reference C output,
it seems that a HIR to LIR promotion does not occur if the block is
already at the top of the stack. This occurs when a HIR is accessed
multiple times in a row and added to the stack on a non-resident miss.
I don't see where this is stated in the paper, though.

Most traces now have matching miss counts. A few, like sprite, are off
by one. So this needs to be looked at next.