Bound the number of LIRS non-resident entries
The paper does not discuss bounding the number of non-resident entries
and relies on prunning the bottom for an ad hoc bounding. The author's
reference implementation has a maximum stack length parameter, which is
emulated here. To avoid traversing resident entries a non-resident queue
should be used instead, per Jackrabbit's version.