W-TinyLfu integration (in-progress)
Early support the for Window TinyLfu eviction policy. This integration
breaks a lot of tests, which are disabled to allow a reasonably sized
commit. Most of the remaining issues are assumptions around a single
access-order queue that need to be corrected. Handling of weights
and new policy tests are not yet started.