Improved eviction handling in W-TinyLfu
1. An entry with a weight of zero may be resurrected if mistakenly
chosen for eviction. This can occur due to a race, where the choice
is made concurrently with the update. This correctness allows a cache
to be used in a state machine, e.g. assigning zero weight to an
in-flight transaction. We've seen custom hacks to CLHM to do this.

2. The main space now tolerates zero weights, which can occur when an
entry's weight is changed after it been moved. As in the eden space,
the entry is skipped over as ineligable.

3. Evictions from the MRU position can only occur for newly added eden
space candidates. If there are no more candidates than the victim is
chosen. This ensures that if the weight is exceeded then the LRU entry
isn't evicted.