Cleanup ordered iteration handling
This preps for moving zero weight entries to their own queue to avoid
scanning by the eviction policy. To help find bugs, now the tests also
validate against a zero weigher by default.

The numerous queues make it complex to provide an order, which is
useful in cases like warm restarts (e.g. Cassandra). This messiness
is better by composing smarter iterators instead of doing the work
manually. This is a minor feature and now won't be as frustrating to
retain.