Disable fastpath for hot reads
The idea was that the hottest entries don't need to be reordered on
reads, due to already being in a desirable state. Based on the power
law, this would avoid unnecessary work for the majority of reads. When
the hot entries are demoted, due to colder entries being promoted, the
fast path wouldn't apply and it would regain its status. Thus little
impact to the hit rate for faster throughput.

Unfortunately benchmarks show otherwise. The throughput varies wildly
and is always lower. This might be due to some type of thrashing in the
CPU caches, but I'm really not sure. So its turned off but can be enabled
in the future if a solution is found.