Use spaces (vs object padding) for ring buffers
Padding is used to avoid false sharing of the cache lines. Previously
the bloat of an object (header, etc) was used as filler, which is a
little better on a laptop. On a server-class machine, spacing the value
out in a primitive array is more efficient. So favoring large machine
over small, but the difference is not significant to bother anyone in
practice.