NPE due to improperly scheduled write tasks (fixes #35)
When an absent computation returns null, a add/removal task was being
improperly scheduled. As there was no Node, this null field was then
causing an NPE during the maintenance cycle. Usually that was on a
background executor, but could be visible if `cleanUp` is called.

This wasn't caught by tests since the executor's was swalled, logged,
and logging was disabled to avoid spamming the report. Now the executor
is instrumented so that the validation can assert no failures occurred.
This detected the two bugs found with existing tests, so this provides
more coverage than a one-off fix.